PART OF ADD JWT
i'm doing a issue about JWT token (hehehe sory for not waiting your merge about the issue of encrypt password).
when run, this commit:
1 -  rename the function that returns the encrypt password correspondent of an email registered for (validate_user_email_returning_encrpyt_password).
2 - Add class token in auth.py. this class contains a function that creates a json web token base on user's password and another function that decode a json web token, the parameters of function is a token and the secret key.
3 - create a login function in api.py that function call another functions called (login_logica) localizated in use_cases_turismo.py
4 - create a functions called: login_logica() and token_required().
login_logica(): verify the login datas and returns a token.
token_required(): this function ensures that all funtions that need login are accesseds only with a valid token.