correcting the past commit - Altering return of repository functions
When run, this commit corrects the previous. In the previous, i stored the value of connection.commit thinking that it would contain the data that was entered in database, but the connection.commit return None. So i refactor this functions inserting a  database query selecting doing a search and find the respective data that has just been stored.